Update README.md
This is a cinema reservation app made with dart & flutter. You can add a movie in the cloustore database, you can perform CRUD operations on the movies made available. You can add or delete movies in a favourites or 'watch later' list. You can view trailers, movie ratings and reviews and as well as make your own. You can check a movie's details, screen times, purchase seats, and receive details of purchases. I started learning dart and flutter in January 2019 and by March I started writing code for this app. Again, it is not the most elegant, or even the least, but maybe it can help new developers avoid some of the mistakes I made. I am still making flutter apps and learning more on the framework. More apps will be uploaded in the future. Thanks!